HSFL League Bot/requirements.txt
disnake==2.9.1
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
fuzzywuzzy>=0.18.0
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0
requests>=2.28.0
unidecode>=1.3.0
python-dotenv>=1.0.0
